fix: refactor castor, everything did related is now into the castor folder Signed-off-by: Francisco Javier Ribo Labrador <elribonazo@gmail.com>
